Made the pi and arduino work together.
There seems to be a known issue where, on serial open the arduino
will end up resetting. This causes some state issues with my code,
so delays had to be put in. A delay was also added to give the arduino
to finish any animations, before closing the serial and we can
continue.

Signed-off-by: Tom Schutte <tjschutte@wisc.edu>